moved spotify api nteraction into login interactor
