+++ src/app/Main.java
@@ -1,6 +1,6 @@
 package app;
 
-import data_access.login.SpotifyLoginDataAccessObject;
+import data_access.UserDataAccessObject;
 import interface_adapter.ViewManagerModel;
 import interface_adapter.logged_in.LoggedInViewModel;
 import interface_adapter.login.LoginViewModel;
@@ -32,7 +32,7 @@ public class Main {
 
         LoginDataAccessInterface userDataAccessObject;
 
-        userDataAccessObject = new SpotifyLoginDataAccessObject("./user.csv");
+        userDataAccessObject = new UserDataAccessObject("./user.csv");
 
         LoginView loginView =  LoginUseCaseFactory.create(viewManagerModel,loginViewModel, loggedInViewModel, userDataAccessObject);
         views.add(loginView, loginView.viewName);
