+++ user.csv
@@ -1,0 +1,2 @@
+name,access_token,login_time
+{  "display_name" : "Anubha",  "external_urls" : {    "spotify" : "https://open.spotify.com/user/6vfmi8iar64pxs04xbbh2j6ri"  },  "href" : "https://api.spotify.com/v1/users/6vfmi8iar64pxs04xbbh2j6ri",  "id" : "6vfmi8iar64pxs04xbbh2j6ri",  "images" : [ ],  "type" : "user",  "uri" : "spotify:user:6vfmi8iar64pxs04xbbh2j6ri",  "followers" : {    "href" : null,    "total" : 3  },  "country" : "CA",  "product" : "premium",  "explicit_content" : {    "filter_enabled" : false,    "filter_locked" : false  },  "email" : "anubhajoshi01@gmail.com"},BQD-2EjjCcdAFgeW_E0pwj9WxSc-YV-iNZpIuDbU21Te0dmiKSCe2KE3YdVXsq0K44ucsjz7EgseE3G3U8fUf_MCD8jGI6z5BVncq0-HE_MVO0Y0jqVAsTb_x3Bvza5Q1kiPXckgZkgCCtq1iTfSrmhBjq7CXWn_ZPnLHmV6_Oxu4vWCXUFKPjquVPVQIQWPw_2s13NmJtc28F-iZuUk_g,16:25:37.609425
